+++ src/processor/operator/order_by/key_block_merger.cpp
@@ -299,7 +299,7 @@ void KeyBlockMergeTaskDispatcher::doneMorsel(std::unique_ptr<KeyBlockMergeMorsel
 void KeyBlockMergeTaskDispatcher::init(MemoryManager* memoryManager,
     std::shared_ptr<std::queue<std::shared_ptr<MergedKeyBlocks>>> sortedKeyBlocks,
     std::vector<std::shared_ptr<FactorizedTable>>& factorizedTables,
-    std::vector<StrKeyColInfo>& strKeyColsInfo, uint64_t numBytesPerTuple) {
+    std::vector<StrKeyColInfo>& strKeyColsInfo, uint32_t numBytesPerTuple) {
     assert(this->keyBlockMerger == nullptr);
     this->memoryManager = memoryManager;
     this->sortedKeyBlocks = sortedKeyBlocks;
